Refactor file naming in `api.py` to simplify output filenames, enhance logging in `anotate_interactors.py` for better traceability during processing, and update `intact_uniprot_pipeline` to allow customizable file patterns for input files. Additionally, improve partner ID isolation logic in `isolate_interactors.py` to exclude certain ID formats.

+++ pipeline/api_intactUNI.py
@@ -8,7 +8,7 @@ if str(parent_dir) not in sys.path:
 
 from intAct.api import fetch_intact_interactions
 
-def intact_uniprot_pipeline(ids_dir, out_dir):
+def intact_uniprot_pipeline(ids_dir, out_dir, file_pattern="_uniprot_ids_COMBINED"):
     """
     For every file in ids_dir, treat the filename (without '_uniprot_ids_COMBINED.txt') as a gene,
     fetch IntAct interactions for each UniProt ID listed, and save results in out_dir.
@@ -17,8 +17,10 @@ def intact_uniprot_pipeline(ids_dir, out_dir):
     out_dir = Path(out_dir)
     out_dir.mkdir(exist_ok=True)
 
-    for file in ids_dir.glob("*_uniprot_ids_COMBINED.txt"):
-        gene_name = file.stem.replace("_uniprot_ids_COMBINED", "")
+    
+    
+    for file in ids_dir.glob('*'+file_pattern+".txt"):
+        gene_name = file.stem.replace(file_pattern, "")
 
         with open(file, 'r') as f:
             uniprot_ids = [line.strip() for line in f if line.strip()]
@@ -37,12 +39,13 @@ def intact_uniprot_pipeline(ids_dir, out_dir):
             df = pd.read_csv(file, sep="\t")
             all_results.append(df)
             files_to_remove.append(file)
-        combined_file = out_dir / f"intact_{gene_name}_UNIPROT.tsv"
-        pd.concat(all_results).to_csv(combined_file, sep="\t", index=False)
-  
-        for file in files_to_remove:
-            if file != combined_file:
-                file.unlink()
+        if all_results:
+            combined_file = out_dir / f"intact_{gene_name}_UNIPROT.tsv"
+            pd.concat(all_results).to_csv(combined_file, sep="\t", index=False)
+    
+            for file in files_to_remove:
+                if file != combined_file:
+                    file.unlink()
 
 
 def main():
